requirements.txt
fastapi>=0.110
uvicorn>=0.23
rich>=13.7
pyyaml>=6.0
httpx>=0.27
websockets>=12.0
python-dotenv>=1.0
